refactor: category MVC 재정의
